pyproject.toml
[project]
name = "polla-scraper"
version = "3.0.0"
description = "HTTP-based scraper for Chilean lottery transparency using alternative data sources"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "click>=8.1.7",
    "requests>=2.32.4",
    "tenacity>=8.5.0",
    "google-auth>=2.35.0",
    "google-api-python-client>=2.146.0",
]

[project.scripts]
polla = "polla_app.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
